+++ src/conponents/TodoList.tsx
@@ -1,0 +1,25 @@
+import { Stack, StackDivider } from '@chakra-ui/react';
+import React, { Dispatch, SetStateAction } from 'react';
+
+import Todo from './Todo';
+
+type Todo = {
+  title: string;
+};
+
+type Props = {
+  todos: Todo[];
+  setTodos: Dispatch<SetStateAction<Todo[]>>;
+};
+
+const TodoList = ({ todos, setTodos }: Props) => {
+  return (
+    <Stack divider={<StackDivider borderColor="gray.200" />} spacing={4}>
+      {todos.map((todo, i) => (
+        <Todo key={todo.title} index={i} title={todo.title} todos={todos} setTodos={setTodos} />
+      ))}
+    </Stack>
+  );
+};
+
+export default TodoList;
